+++ src/Exception/InvalidTokenException.php
@@ -7,12 +7,4 @@ namespace Abb\Fakturownia\Exception;
  */
 class InvalidTokenException extends \InvalidArgumentException
 {
-
-    /**
-     * Constructor
-     */
-    public function __construct()
-    {
-        parent::__construct('Invalid Fakturownia API token');
-    }
 }