+++ src/FakturowniaTokenValidator.php
@@ -21,11 +21,13 @@ class FakturowniaTokenValidator
      */
     public function isValidTokenOrFail($token): void
     {
-        // check pattern: token_hash/username
+        // check pattern: token_hash/subdomain
         $isValid = is_string($token) && 1 === preg_match('~^[^/]+/[^/]+$~', $token);
 
         if (!$isValid) {
-            throw new InvalidTokenException();
+            throw new InvalidTokenException(
+                'Invalid API token. Only tokens with prefix are supported. You can generate it in fakturownia.pl service.'
+            );
         }
     }
 }